*: move pkg/ansible to internal/ansible
