+++ .gitignore
@@ -101,7 +101,7 @@ images/scorecard-test/scorecard-test
 images/scorecard-test-kuttl/scorecard-test-kuttl
 
 # Test artifacts
-pkg/ansible/runner/testdata/valid.yaml
+internal/ansible/runner/testdata/valid.yaml
 /bin
 internal/scorecard/examples/custom-scorecard-tests/go.sum
