+++ changelog/fragments/mv-ansible-internal.yaml
@@ -1,0 +1,11 @@
+entries:
+  - description: >
+      The ansible operator implementation in `pkg/ansible` was moved to
+      `internal/ansible`.
+
+    kind: change
+    breaking: true
+    migration:
+      header: Hybrid Go/Ansible operator use cases are not supported in 1.0.0
+      body: >
+        There is no migration path that enables continued use of the Ansible-based operator Go libraries
