+++ cmd/ansible-operator/main.go
@@ -32,14 +32,14 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/manager/signals"
 
+	"github.com/operator-framework/operator-sdk/internal/ansible/controller"
+	"github.com/operator-framework/operator-sdk/internal/ansible/flags"
+	"github.com/operator-framework/operator-sdk/internal/ansible/proxy"
+	"github.com/operator-framework/operator-sdk/internal/ansible/proxy/controllermap"
+	"github.com/operator-framework/operator-sdk/internal/ansible/runner"
+	"github.com/operator-framework/operator-sdk/internal/ansible/watches"
 	"github.com/operator-framework/operator-sdk/internal/log/zap"
 	"github.com/operator-framework/operator-sdk/internal/util/k8sutil"
-	"github.com/operator-framework/operator-sdk/pkg/ansible/controller"
-	"github.com/operator-framework/operator-sdk/pkg/ansible/flags"
-	"github.com/operator-framework/operator-sdk/pkg/ansible/proxy"
-	"github.com/operator-framework/operator-sdk/pkg/ansible/proxy/controllermap"
-	"github.com/operator-framework/operator-sdk/pkg/ansible/runner"
-	"github.com/operator-framework/operator-sdk/pkg/ansible/watches"
 	sdkVersion "github.com/operator-framework/operator-sdk/version"
 )
