+++ internal/ansible/controller/controller.go
@@ -32,9 +32,9 @@ import (
 	ctrlpredicate "sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/source"
 
-	"github.com/operator-framework/operator-sdk/pkg/ansible/events"
-	"github.com/operator-framework/operator-sdk/pkg/ansible/predicate"
-	"github.com/operator-framework/operator-sdk/pkg/ansible/runner"
+	"github.com/operator-framework/operator-sdk/internal/ansible/events"
+	"github.com/operator-framework/operator-sdk/internal/ansible/predicate"
+	"github.com/operator-framework/operator-sdk/internal/ansible/runner"
 )
 
 var log = logf.Log.WithName("ansible-controller")
