+++ internal/ansible/predicate/predicate.go
