+++ internal/ansible/runner/internal/inputdir/inputdir.go
