+++ internal/ansible/runner/runner_test.go
@@ -24,7 +24,7 @@ import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 
-	"github.com/operator-framework/operator-sdk/pkg/ansible/watches"
+	"github.com/operator-framework/operator-sdk/internal/ansible/watches"
 )
 
 func checkCmdFunc(t *testing.T, cmdFunc cmdFuncType, playbook, role string, verbosity int) {
