+++ internal/ansible/watches/testdata/duplicate_gvk.yaml
