+++ internal/ansible/watches/testdata/invalid.yaml
