+++ internal/ansible/watches/testdata/invalid_collection.yaml
