+++ internal/ansible/watches/testdata/invalid_finalizer_no_vars.yaml
