+++ internal/ansible/watches/testdata/invalid_finalizer_playbook_path.yaml
