+++ internal/ansible/watches/testdata/invalid_finalizer_role_path.yaml
