+++ internal/ansible/watches/testdata/invalid_role_path.yaml
