+++ internal/ansible/watches/testdata/invalid_status.yaml
