+++ internal/ansible/watches/testdata/invalid_yaml_file.yaml
