+++ internal/ansible/watches/testdata/valid.yaml.tmpl
